requirements.txt
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
charset-normalizer==3.4.7
chess==1.11.2
click==8.3.2
contourpy==1.3.3
cuda-bindings==13.2.0
cuda-pathfinder==1.5.3
cuda-toolkit==13.0.2
cycler==0.12.1
fastapi==0.135.3
filelock==3.28.0
fonttools==4.62.1
fsspec==2026.3.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
idna==3.11
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
narwhals==2.19.0
networkx==3.6.1
numpy==2.4.4
nvidia-cublas==13.1.0.3
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
opencv-python==4.13.0.92
packaging==26.1
pandas==3.0.2
pillow==12.2.0
polars==1.39.3
polars-runtime-32==1.39.3
protobuf==7.34.1
psutil==7.2.2
pyarrow==23.0.1
pydantic==2.13.0
pydantic_core==2.46.0
pydeck==0.9.1
pydobot==1.3.2
pyparsing==3.3.2
pyserial==3.4
python-chess==1.999
python-dateutil==2.9.0.post0
PyYAML==6.0.3
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
scipy==1.17.1
setuptools==81.0.0
six==1.17.0
smmap==5.0.3
starlette==1.0.0
streamlit==1.56.0
sympy==1.14.0
tenacity==9.1.4
toml==0.10.2
torch==2.11.0
torchvision==0.26.0
tornado==6.5.5
triton==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.4.37
ultralytics-thop==2.0.18
urllib3==2.6.3
uvicorn==0.44.0
watchdog==6.0.0
